yolo-service/requirements.txt
Flask
Flask-CORS
opencv-python-headless
torch
pika
deep_sort_realtime
python-dotenv
